src/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1.2
django-bootstrap4==22.3
django-crispy-forms==1.14.0
django-filter==22.1
django-template-maths==0.2.0
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.5
uuid==1.30
whitenoise==6.2.0
